requirements.txt
supabase
fastapi-mcp
uvicorn
pydantic[email]
python-jose
passlib
openai
